feat(examples): add blaze to hello world
