+++ examples/hello_world/src/generate-credentials.ts
